==> 0_setup.ipynb <==
# Let's mount our G-Drive. 

from google.colab import drive
drive.mount('/content/drive')

# # Create the Workspace folder
!mkdir -p '/content/drive/My Drive/ICDMAI_Tutorial/notebook/'

# # Clone the repository
!git clone 'https://github.com/amitbcp/icdmai_2020.git' '/content/drive/My Drive/ICDMAI_Tutorial/notebook/'

# Download the Data
!gdown 'https://drive.google.com/uc?id=1XTkF0yANUPIZ3SXw5BbbWKCkWo7ZK56T'

# Unzip the Data
!unzip 'ICDMAI_Tutorial.zip' -d '/content/drive/My Drive/'

# Annnd You are ready to Go !

## Other Artificats

To run End-to-End pipeline you would require more data & space in G-Drive. If you are sure of it go-forward & run the cells below.

# Download the raw-data files
!gdown 'https://drive.google.com/uc?id=1gkgUlkaRXUzrNR_jY42ieK4xtLX3ztKX'
!unzip 'raw_data_files.zip' -d '/content/drive/My Drive/ICDMAI_Tutorial/'

# Download the data for Single Corpus & Evaluation
!gdown 'https://drive.google.com/uc?id=1csNEFbKMxzQ8FQhudJ9K1Bn9SnRLyAug'
!unzip '70_30_split.zip' -d '/content/drive/My Drive/ICDMAI_Tutorial/notebook/training_data/'
