==> 2_classical_ml_approach.ipynb <==
# Classical Machine Learning Approach

In this notebook we will be learning to
  1. Create a Naive TF - IDF based Bag of Words representation of text.
  2. Use classical ML models to solve text classification.
  3. Use a One Vs Rest strategy to solve multi-label text classification.


  **HOT TIP** : *Save them as pickle for easy rendering for experiments*

  This Notebook uses code from https://github.com/susanli2016/Machine-Learning-with-Python/blob/master/Multi%20label%20text%20classification.ipynb


# Installing packages.
!pip install contractions
!pip install textsearch
!pip install tqdm

# Importing packages.
import nltk
nltk.download('punkt')
nltk.download('stopwords')
%matplotlib inline
import re
import matplotlib
import numpy as np
import matplotlib.pyplot as plt
import pandas as pd
from sklearn.model_selection import train_test_split
from sklearn.feature_extraction.text import TfidfVectorizer
from sklearn.naive_bayes import MultinomialNB
from sklearn.metrics import accuracy_score
from sklearn.multiclass import OneVsRestClassifier
from nltk.corpus import stopwords
stop_words = set(stopwords.words('english'))
from sklearn.svm import LinearSVC
from sklearn.linear_model import LogisticRegression
from sklearn.pipeline import Pipeline
import seaborn as sns
from sklearn.metrics import confusion_matrix, classification_report
import pickle
import ast
from sklearn.externals import joblib
from datetime import datetime
from sklearn.preprocessing import MultiLabelBinarizer

# Let's mount our G-Drive.

from google.colab import drive
drive.mount('/content/drive', force_remount=True)

# Data read and preparation.
# Mentioning where is our data located on G-Drive. Make sure to rectify your path
path = '/content/drive/My Drive/ICDMAI_Tutorial/notebook/'
data ='filtered_data/question_tag_text_mapping.pkl'
ml_model = path + 'ml_model/'

# Let us quickly load our question tag data
question_tag = pd.read_pickle(path+data)
question_tag.head(3)

### Creating one hot encoding from multilabelled tagged data

# In order to use one vs rest strategy we will need to one hot encoding each tag across all documents.
mlb = MultiLabelBinarizer()
question_tag['Tag_pop'] = question_tag['Tag']
question_tag = question_tag.join(pd.DataFrame(mlb.fit_transform(question_tag.pop('Tag_pop')),
                          columns=mlb.classes_,
                          index=question_tag.index))
question_tag.head(3)

# Creating a list of all existing 'Tags'
dummy = question_tag.drop(['Id', 'OwnerUserId', 'CreationDate', 'ClosedDate', 'Score', 'Title','Body','Tag'], axis=1)
categories = list(dummy.columns.values)

### Text preprocessing

# Let us createa a very basic text preprocessor which we will use for cleaning text.
def clean_text(text):
    text = text.lower()
    text = re.sub(r"what's", "what is ", text)
    text = re.sub(r"\'s", " ", text)
    text = re.sub(r"\'ve", " have ", text)
    text = re.sub(r"can't", "can not ", text)
    text = re.sub(r"n't", " not ", text)
    text = re.sub(r"i'm", "i am ", text)
    text = re.sub(r"\'re", " are ", text)
    text = re.sub(r"\'d", " would ", text)
    text = re.sub(r"\'ll", " will ", text)
    text = re.sub(r"\'scuse", " excuse ", text)
    text = re.sub('\W', ' ', text)
    text = re.sub('\s+', ' ', text)
    text = text.strip(' ')
    return text

question_tag['Body'] = question_tag['Body'].map(lambda com : clean_text(com))

### Creating a 70/30 Train-Test Split

train, test = train_test_split(question_tag, random_state=42, test_size=0.30, shuffle=True)

X_train = train.Body
X_test = test.Body

print("Train data shape : {}".format(X_train.shape))
print("Test data shape : {}".format(X_test.shape))

# Creating Bag of Words representation using TF - IDF
  1. Initializing the Vectorizer object
  2. Create a corpus from training data.
  3. Create a document term matrix

#Initializing the Vectorizer object
tfidf = TfidfVectorizer(stop_words=stop_words)

#Create a corpus from training data
#Create a document term matrix of training data based on the corpus.
X_train_dtm = tfidf.fit_transform(X_train)

#Create a document term matrix of test data based on the corpus.
#Note that the dimensions/columns of DTM of the test data will be based on the training data corpus only.
X_test_dtm = tfidf.transform(X_test)

## Pipeline
scikit-learn provides a Pipeline utility to help automate machine learning workflows. Pipelines are very common in Machine Learning systems, since there is a lot of data to manipulate and many data transformations to apply. So we will utilize pipeline to train every classifier.

## One Vs Rest Multilabel strategy
The Multi-label algorithm accepts a binary mask over multiple labels. The result for each prediction will be an array of 0s and 1s marking which class labels apply to each row input sample.

OneVsRest strategy can be used for multilabel learning, where a classifier is used to predict multiple labels for instance. **Naive Bayes**, **SVM**, **Logistic Regression** supports multi-class, but we are in a multi-label scenario, therefore, we wrap them in the OneVsRestClassifier.

### We create a Training Pipeline and a Scoring Pipeline

def tag_level_training_pipeline(X_train, train, X_test, test, classifier_pipeline, output_directory):
  
  #1. Create a classifier for each Tag
  for category in categories:
    print('... Processing {}'.format(category))
    
    # 1. train the model using X_dtm & y
    classifier_pipeline.fit(X_train, train[category])
    
    # 2. save the model to disk
    filename = ml_model + output_directory +str(category)+ '_model.pkl'
    joblib.dump(classifier_pipeline, filename, compress = 1)
    
    # 3. compute the testing accuracy
    prediction = classifier_pipeline.predict(X_test)
    print('Test accuracy is {}'.format(accuracy_score(test[category], prediction)))
    print(classification_report(test[category], prediction))



def tag_level_predict(X_train, train, X_test, test, model_directory):
  prediction_df = pd.DataFrame(columns=['dummy1'])
  
  #Score the document across classifier for each Tag
  for category in categories:
    
    # 1. load the model
    filename = ml_model + model_directory +str(category)+ '_model.pkl'
    classifier_pipeline = joblib.load(filename)
    
    # 2. predict on the test data.
    prediction = classifier_pipeline.predict(X_test)
    prediction_df[str(category)] = prediction

  # Remember We had encoded the labels. It time to bring them back to their original form.
  for category in categories:
    prediction_df.loc[prediction_df[str(category)] == 1, str(category)] = category
  prediction_df['predicted_labels'] = prediction_df[[str(i) for i in categories]].values.tolist()
  prediction_df['predicted_labels'] =  prediction_df['predicted_labels'].apply(lambda x : list(set(x)))
  # prediction_df['predicted_labels'] = prediction_df['predicted_labels'].apply(lambda x: x.remove(0) if (0 in x) else x )
  
  # We create result having orignal labels and predicted labels for metrics Evaluation
  final_pred_df = pd.concat([test[['Id','Tag']].reset_index(), prediction_df[['predicted_labels']].reset_index()], axis=1)
  final_pred_df['original_labels'] = final_pred_df['Tag']
  # prediction_df[['Id']] = test[['Id']]
  final_pred_df_result = final_pred_df[['Id','original_labels','predicted_labels']]
  return final_pred_df_result

# importing os module 
import os
try:
  os.rename('/content/drive/My Drive/ICDMAI_Tutorial/notebook/ml_model/SVM/_net_model.pkl', '/content/drive/My Drive/ICDMAI_Tutorial/notebook/ml_model/SVM/.net_model.pkl')
except :
  print("Already in proper filename!")  

## A Dummy example.
X_test = ["How to handle memory locking ?", "How to handle memory locking in java ?", "How to handle memory locking in java python ?","This post is not about java"]
X_test_dtm = tfidf.transform(X_test)
result = tag_level_predict(X_train_dtm, train, X_test_dtm, test.head(1), 'SVM/')

for i in range(result.shape[0]):
  print("Input [",X_test[i],"] || Predicted classes: ",result.predicted_labels[i])

# Evaluating our results

# Here we define precision, recall, f1 measure at a single document level.
def document_evaluation_metrics(prd_grp,grp,metric="precision"):
    pred_group = prd_grp
    if 0 in pred_group: pred_group.remove(0)
    group = grp

    set_pred_group = set(pred_group)
    set_group = set(group)
    intrsct = set_group.intersection(set_pred_group)
    accuracy = len(intrsct) / float(len(set_pred_group) if len(set_pred_group)>1 else 1)
    recall = len(intrsct) / float(len(set_group) if len(set_group)>1 else 1)
    if metric == "precision":
      return accuracy
    elif metric == "recall":
      return recall
    elif metric == "f1_measure":
      if accuracy == 0 or recall == 0:
        return 0
      elif accuracy > 0 and recall >0 :
        f1_measure = 2*accuracy*recall/(float(accuracy + recall))
        return f1_measure
    
    return -1

# Provide overall average stats and populate document level metrics.
def model_evaluation_stats(final_pred_df, model_name="default"):
  final_pred_df['doc_precision'] = final_pred_df.apply(lambda x: document_evaluation_metrics(x.predicted_labels, x.original_labels, "precision"), axis=1)
  final_pred_df['doc_recall'] = final_pred_df.apply(lambda x: document_evaluation_metrics(x.predicted_labels, x.original_labels, "recall"), axis=1)
  final_pred_df['doc_f1_measure'] = final_pred_df.apply(lambda x: document_evaluation_metrics(x.predicted_labels, x.original_labels, "f1_measure"), axis=1)
  
  print('Avearge precision across documents is {}'.format(final_pred_df['doc_precision'].mean()))
  print('Avearge recall across documents is {}'.format(final_pred_df['doc_recall'].mean()))
  print('Avearge f1 measure across documents is {}'.format(final_pred_df['doc_f1_measure'].mean()))
  pickle.dump(final_pred_df, open(ml_model + model_name + ".pkl", 'wb'))
  # final_pred_df.to_csv(ml_model + 'SVM_Tag_predictions.txt',sep='\t',index=False)

# Let us train, score and evaluate Naive Bayes

#Naive Bayes Classifier
NB_pipeline = Pipeline([
                ('clf', OneVsRestClassifier(MultinomialNB(
                    fit_prior=True, class_prior=None))),
            ])

tag_level_training_pipeline(X_train_dtm, train, X_test_dtm, test, NB_pipeline, 'NaiveBayes/')
result = tag_level_predict(X_train_dtm, train, X_test_dtm, test, 'NaiveBayes/')
model_evaluation_stats(result, "NaiveBayes")

# Let us train, score and evaluate Support Vector Machines

#SVM Classifier
SVC_pipeline = Pipeline([
                ('clf', OneVsRestClassifier(LinearSVC(), n_jobs=1)),
            ])

tag_level_training_pipeline(X_train_dtm, train, X_test_dtm, test, SVC_pipeline, 'SVM/')
result = tag_level_predict(X_train_dtm, train, X_test_dtm, test, 'SVM/')
model_evaluation_stats(result, "SVM")

# Let us train, score and evaluate Logistic Regression

#Logistic Regression Classifier
LogReg_pipeline = Pipeline([
                ('clf', OneVsRestClassifier(LogisticRegression(solver='sag'), n_jobs=1)),
            ])

tag_level_training_pipeline(X_train_dtm, train, X_test_dtm, test, LogReg_pipeline, 'LogisticRegression/')
result = tag_level_predict(X_train_dtm, train, X_test_dtm, test, 'LogisticRegression/')
model_evaluation_stats(result, "LogisticRegression")
