==> 4_word_embedding.ipynb <==
# Create Custom Word Embeddings

In this notebook we aim to create our custom word-embeddings which are generic/unsupervised and can be used in any of our NLP applications where we use a word embeddings.

We majorly will be doing the following :

  1. Install & Import Packages
  2. Define our Text Pre-processing pipeline.
  3. Creating Text Corpus
  4. Train Word Embeddings
  5. Convert & Save 

# Let's mount our G-Drive. 

from google.colab import drive
drive.mount('/content/drive')

# 1. Install & Import Packages

Make sure you restart the runtime once the pip installs has completed.

!pip install -U fasttext
!pip install -U gensim
!pip install tiny-tokenizer
!pip install flair

# Let's import our packages !

import pandas as pd
import fasttext
import re
import html
from tqdm import tqdm 
from flair.embeddings import WordEmbeddings

# 2. Text Pre-processing Pipeline

In NLP with the structured & unstructured text, one constant that doen't change is text pre-processing. Based on the task & requirements we create a text-cleaning pipeline.

Every try-except block can be written as a different modular function which can be invoked from preprocess_text() function. This serves as a pipeline of the series of text-cleaning that you might require for your dataset.

clean = re.compile('<.*?>')

def preprocess_text(text) :
  try :
    # soup = BeautifulSoup(text, "html.parser")
    # text = soup.get_text()
    text=  re.sub(clean, '', text)
    text = html.unescape(text)
  except Exception as e:
    print("Error in HTML Processing ...")
    print(text)
    text = text
    raise e
  try :
    # remove extra newlines (often might be present in really noisy text)
    text = text.translate(text.maketrans("\n\t\r", "   "))
  except :
    print("Error in removing extra lines ...")
    print(text)
    text = text

  try :
    # remove extra whitespace
    text = re.sub(' +', ' ', text)
    text = text.strip()
  except :
    print("Error in extra whitespace removal ...")
    print(text)
    text = text

  return text

# 3. Creating the Corpus

To train a word-embedding we first need to create a single corpus which contains all of our text that we want the `Machines` to understand. This includes the following :

1. Loading all the Documents (Text Files/Sentences/Documents etc) & pre-process
2. *One Line Per Document* : Traning format
3. Save to re-use.
4. Train FastText Embedding
5. Convert to Gensim re-usable & distributable format

# Define the Base Path & Data Files
path = '/content/drive/My Drive/ICDMAI_Tutorial/notebook/'
data ='filtered_data/question_tag_text_mapping.pkl'

# Load the structured file with all document Text
question_tag = pd.read_pickle(path+data)
question_tag.head()

# Apply the pre-processing on the Question Body & Text

question_tag['Title'] = question_tag['Title'].apply(preprocess_text)
question_tag['Body'] = question_tag['Body'].apply(preprocess_text)
question_tag.head()

# Iterate & Create "One Line Per Document : Traning format"

text_lines =  list()
for index in tqdm(question_tag.index) :
  title = question_tag.loc[index,'Title']
  body =  question_tag.loc[index,'Body']

  text =  title + '. ' + body
  text_lines.append(text)

# Save to re-use easily next-time
with open(path+'/training_data/training_data.txt', 'w',encoding ='utf-8') as filehandle:
    filehandle.writelines("%s\n" % sent for sent in text_lines)

# 4. Train Word Embeddings

Here we use FastText to train wor-embeddings as these are very fast compared to any other trainings available and serves the purpose to evaluate how valuable custom embeddings are to the task at hand.


FastText open-sourced by Facebook, provides both command-line based library & python based. If possible use command-line based as it's more faster.

The options & hyperparameters to be tune can be found in the documentation.


**Link**  : https://github.com/facebookresearch/fastText/tree/master/python#train_unsupervised-parameters

# Skipgram model :
model = fasttext.train_unsupervised(path+'/training_data/training_data.txt', model='skipgram',dim=300,verbose=3)
model.save_model(path+"/training_data/training_data_skipgram_d300.bin")
print("Model Saved")

# 5. Convert & Save

Generally you would want to convert the embedding to Gensim format as it's a widely used library and provides wrapper to most framework.

from gensim.models import KeyedVectors

# We use the .vec file which does not contain the Meta-Data of FastText word-embeddings
model = KeyedVectors.load_word2vec_format('/content/drive/My Drive/ICDMAI_Tutorial/notebook/training_data/word_embedding/model.vec')

# Save
model.save('/content/drive/My Drive/ICDMAI_Tutorial/notebook/training_data/word_embedding/gensim_model')

# Test that we are able to load it in for the down-stream tasks/code based we want.
word_embeddings = [ WordEmbeddings('/content/drive/My Drive/ICDMAI_Tutorial/notebook/training_data/word_embedding/gensim_model')]
print("Word Embedding Loaded Successfully in Flair")

# Looking at the embeddings
from flair.embeddings import WordEmbeddings
from flair.data import Sentence
word_embeddings = WordEmbeddings('/content/drive/My Drive/ICDMAI_Tutorial/notebook/training_data/word_embedding/gensim_model')

# some example sentence
sentence = Sentence('Welcome to ICDMAI 2020')
# sentence = Sentence(question_tag.head(1))

# embed sentences
word_embeddings.embed(sentence)

# go through each token in sentence
for token in sentence:
    # print token string
    print("Word : {}".format(token.text))
    # print embedding of this Token
    print(token.embedding)

    # print shape of embedding of this Token
    print(token.embedding.shape)
