==> 6_inference_pipeline.ipynb <==
# Perdiction Pipeline | Ensemble Pipeline

In this notebook, we will ensemble our models to make a prediction pipeline.
The keys steps would be as following :

  1. Text Preprocessing for inference.
  2. Load classifiers iteratively in a list.
  3. Load test-data & pre-process.
  4. Set classifier threshold and run it through Ensemble



# First let's check what has Google given us ! Thank you Google for the GPU

!nvidia-smi

# Let's mount our G-Drive. Hey !! Because for GPU you now give your data to Google 

from google.colab import drive
drive.mount('/content/drive')

# Install necessary packages and restart the environment

!pip install tiny-tokenizer
!pip install  flair

# Let's import our packages !

import pandas as pd
from tqdm import tqdm
import html
import re
from bs4 import BeautifulSoup
import re
from sklearn.model_selection import train_test_split
# import flair
import pickle
from torch.optim.adam import Adam

# The sentence objects holds a sentence that we may want to embed or tag
from flair.data import Sentence
from flair.data import Corpus
from flair.datasets import ClassificationCorpus
from flair.embeddings import WordEmbeddings, FlairEmbeddings, DocumentRNNEmbeddings
from flair.models import TextClassifier
from flair.trainers import ModelTrainer
from flair.samplers import ImbalancedClassificationDatasetSampler

# 1. Text Pre-processing

### Note -
We need to be careful to apply the same set of text transformations during inference as we did during training. Any changes will directly affect the model to produce junk / more junky results

clean = re.compile('<.*?>')

def preprocess_text(text) :
  try :
    # soup = BeautifulSoup(text, "html.parser")
    # text = soup.get_text()
    text=  re.sub(clean, '', text)
    text = html.unescape(text)
  except :
    print("Error in HTML Processing ...")
    print(text)
    text = text
  try :
    # remove extra newlines (often might be present in really noisy text)
    text = text.translate(text.maketrans("\n\t\r", "   "))
  except :
    print("Error in removing extra lines ...")
    print(text)
    text = text

  try :
    # remove extra whitespace
    text = re.sub(' +', ' ', text)
    text = text.strip()
  except :
    print("Error in extra whitespace removal ...")
    print(text)
    text = text

  return text

# 2. Load Classifiers

We iterate through the saved classifiers and load them. We append them in a list can be easily iterated during evaluation/prediction.

classifiers  = []

for grp_id in tqdm(range(1,15)) :
  path  = '/content/drive/My Drive/ICDMAI_Tutorial/notebook/training_data/normalised/group/' + str(grp_id) + '/best-model.pt'
  classifier = TextClassifier.load(path)
  classifiers.append(classifier)

# 3. Load Test Data & Preprocess

We load the test-data from the previously created split and re-transform it into a Dataframe with corresponding labels.

Finally we save the DataFrame with a placeholder for the predicted labels.

test_data = '/content/drive/My Drive/ICDMAI_Tutorial/notebook/training_data/70_30_split/test.txt'

with open(test_data,'r',encoding='utf-8') as f :
  data = f.readlines()

# prefix from the Training Data Format
prefix = '__label__'

test_text = []
test_label = []
for doc in tqdm(data) :
  splits =  doc.split()
  labels = []
  idx = 0 
  for word in splits :
    if prefix in word :
      labels.append(word[9:].strip())
      idx += len(word)
    else :
      text = doc[idx:].strip()
      break
  test_text.append(text)
  test_label.append(labels)
  # break

test_df = pd.DataFrame(list(zip(test_text,test_label)),columns = ['text','original_labels'])
test_df['predicted_labels'] = None


test_df.to_pickle('/content/drive/My Drive/ICDMAI_Tutorial/notebook/training_data/70_30_split/test.pkl')
test_df.head()

# 4. Run the Ensemble Pipeline

Once we load the test-data, we can iteratively run it via all the classifiers and store the predictions for evaluating the performance.

### Coding Exercise 
  1. Multi-thread/process the prediction pipeline.
  2. Find a redundant line of code in the cell below
  3. Optimize it

def predict_ensemble(sentence,threshold=0.1) :
  labels  = [] 
  ## Iterate through each classifier or prediction
  for classifier in classifiers:
    classifier.multi_label_threshold = threshold
    classifier.predict(sentence)
    for label in sentence.labels :
      ## Append labels from all classifiers
      labels.append(label.value)
  return labels

# Read the cleaned Test Data for running with different classifiers & threshold

with open('/content/drive/My Drive/ICDMAI_Tutorial/notebook/training_data/70_30_split/test.pkl','rb') as f :
  test_df = pickle.load(f)

## Setting a global threshold for all classifiers
threshold = 0.9

## Iterating through each Test Example
for idx in tqdm(test_df.index) :
  text = preprocess_text(test_df.loc[idx,'text'])
  
  # create example sentence
  sentence = Sentence(text)
  labels = []

  ## Iterate through each classifier or prediction
  for classifier in classifiers:
    classifier.multi_label_threshold = threshold
    classifier.predict(sentence)

    for label in sentence.labels :
      ## Append labels from all classifiers
      labels.append(label.value)

  test_df.at[idx,'predicted_labels'] = labels


# Save the predictions
test_df.to_pickle('/content/drive/My Drive/ICDMAI_Tutorial/notebook/training_data/70_30_split/normalised_test_predcted_th_0'+ str(int(threshold*10)) +'.pkl')
test_df.head()

# Let's Infer !

Let's manually create some text/questions and ask the model about it.

text1 = preprocess_text("How to handle memory locking ?")
# create example sentence & tokenize
sentence1 = Sentence(text1)
# predict
labels = predict_ensemble(sentence1,threshold=0.8)
print(labels)


text2 = preprocess_text("How to handle memory locking in java ?")
# create example sentence & tokenize
sentence2 = Sentence(text2)
# predict
labels = predict_ensemble(sentence2,threshold=0.8)
print(labels)

text3 = preprocess_text("How to handle memory locking in java python ?")
# create example sentence & tokenize
sentence3 = Sentence(text3)
# predict
labels = predict_ensemble(sentence3,threshold=0.8)
print(labels)

text4 = preprocess_text("This post is not about java")
# create example sentence & tokenize
sentence4 = Sentence(text4)
# predict
labels = predict_ensemble(sentence4,threshold=0.8)
print(labels)
